Create new folders to better manage the different classes, updated runtest.sh
