--- src/main/java/Deadline.java
@@ -1,13 +1,0 @@
-public class Deadline extends Task {
-    protected String deadline;
-
-    public Deadline(String description, String deadline) {
-        super(description);
-        this.deadline = deadline;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + deadline + " )";
-    }
-}
