+++ src/main/java/tasks/Deadline.java
@@ -1,0 +1,15 @@
+package pawpal.tasks;
+
+public class Deadline extends Task {
+    protected String deadline;
+
+    public Deadline(String description, String deadline) {
+        super(description);
+        this.deadline = deadline;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + deadline + " )";
+    }
+}
