+++ src/main/java/tasks/ToDo.java
@@ -1,0 +1,12 @@
+package pawpal.tasks;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
