+++ src/main/java/utils/Command.java
@@ -1,0 +1,3 @@
+public enum Command {
+    LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, INVALID
+}
